[mod] gradle multi module 구조 변경
