+++ api/build.gradle
@@ -1,0 +1,19 @@
+dependencies {
+    implementation project(':common')
+    implementation project(':member')
+    implementation project(':payment')
+    implementation project(':order')
+    implementation project(':product')
+    implementation project(':stock')
+
+    implementation 'org.springframework.boot:spring-boot-starter-security'
+    testImplementation 'org.springframework.security:spring-security-test'
+
+    annotationProcessor 'org.springframework.boot:spring-boot-configuration-processor'
+
+    implementation 'io.jsonwebtoken:jjwt-api:0.12.5'
+    runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.12.5'
+    runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.12.5'
+}
+
+
