+++ api/src/main/java/com/mung/api/config/BeanConfig.java
@@ -1,0 +1,12 @@
+package com.mung.api.config;
+
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@EnableAutoConfiguration
+@ComponentScan(basePackages = {"com.mung.member"})
+public class BeanConfig {
+
+}
