+++ api/src/main/java/com/mung/api/config/jwt/JwtAuthenticationFilter.java
@@ -1,0 +1,59 @@
+package com.mung.api.config.jwt;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.mung.api.config.auth.PrincipalDetails;
+import com.mung.member.domain.Member;
+import io.jsonwebtoken.Jwts;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+
+import java.io.IOException;
+import java.util.Date;
+
+@Slf4j
+@RequiredArgsConstructor
+public class JwtAuthenticationFilter extends UsernamePasswordAuthenticationFilter {
+
+    private final AuthenticationManager authenticationManager;
+
+    private final JwtConfig jwtConfig;
+
+    @Override
+    public Authentication attemptAuthentication(HttpServletRequest request, HttpServletResponse response) throws AuthenticationException {
+
+        try {
+            ObjectMapper objectMapper = new ObjectMapper();
+            Member member = objectMapper.readValue(request.getInputStream(), Member.class);
+            UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken(member.getEmail(), member.getPassword());
+
+            return authenticationManager.authenticate(authenticationToken);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        return null;
+    }
+
+    @Override
+    protected void successfulAuthentication(HttpServletRequest request, HttpServletResponse response, FilterChain chain, Authentication authResult) throws IOException, ServletException {
+        PrincipalDetails principalDetails = (PrincipalDetails) authResult.getPrincipal();
+        
+        String jws = Jwts.builder()
+                .subject("jwt")
+                .claim("id", principalDetails.getMemberId())
+                .expiration(new Date(System.currentTimeMillis() + (1000 * 60 * 10))) // 10분
+                .signWith(jwtConfig.getKeyFromBase64EncodedKey(jwtConfig.key))
+                .compact();
+
+        response.addHeader("Authorization", "Bearer " + jws);
+    }
+
+}
