+++ api/src/main/java/com/mung/api/config/jwt/JwtConfig.java
@@ -1,0 +1,26 @@
+package com.mung.api.config.jwt;
+
+import io.jsonwebtoken.io.Decoders;
+import io.jsonwebtoken.io.Encoders;
+import io.jsonwebtoken.security.Keys;
+import org.springframework.context.annotation.Configuration;
+
+import javax.crypto.SecretKey;
+import java.nio.charset.StandardCharsets;
+
+@Configuration
+public class JwtConfig {
+
+    public String key = "/VLfdrwEn0MZScSqq6En7NMOW9rF7SQHN7LiWGCG21g=";
+
+    public String encodeBase64SecretKey(String secretKey) {
+        return Encoders.BASE64.encode(secretKey.getBytes(StandardCharsets.UTF_8));
+    }
+
+    public SecretKey getKeyFromBase64EncodedKey(String base64EncodedSecretKey) {
+        byte[] keyBytes = Decoders.BASE64.decode(base64EncodedSecretKey);
+
+        return Keys.hmacShaKeyFor(keyBytes);
+    }
+
+}
