+++ common/build.gradle
@@ -1,10 +1,7 @@
 dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-security'
-    testImplementation 'org.springframework.security:spring-security-test'
 
-    annotationProcessor 'org.springframework.boot:spring-boot-configuration-processor'
+}
 
-    implementation 'io.jsonwebtoken:jjwt-api:0.12.5'
-    runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.12.5'
-    runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.12.5'
+bootJar {
+    enabled = false
 }
