--- common/src/main/java/com/mung/common/config/auth/PrincipalDetailsService.java
@@ -1,24 +1,0 @@
-package com.mung.common.config.auth;
-
-import com.mung.common.domain.Member;
-import com.mung.common.repository.LoginRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class PrincipalDetailsService implements UserDetailsService {
-
-    private final LoginRepository loginRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        Member member =  loginRepository.findByEmail(username)
-                .orElseThrow(() -> new UsernameNotFoundException(username + "을 찾을 수 없습니다."));
-
-        return new PrincipalDetails(member);
-    }
-}
