--- common/src/main/java/com/mung/common/config/handler/Http401Handler.java
@@ -1,34 +1,0 @@
-package com.mung.common.config.handler;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.mung.common.response.ErrorResponse;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-
-@Slf4j
-@RequiredArgsConstructor
-public class Http401Handler implements AuthenticationEntryPoint {
-
-    private final ObjectMapper objectMapper;
-
-    @Override
-    public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException) throws IOException, ServletException {
-        log.error("[인증오류] 로그인이 필요합니다.");
-        ErrorResponse errorResponse = ErrorResponse.builder()
-                .code("401")
-                .message("로그인이 필요합니다.")
-                .build();
-        response.setCharacterEncoding(StandardCharsets.UTF_8.displayName());
-        response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-        objectMapper.writeValue(response.getWriter(), errorResponse);
-    }
-
-}
