--- common/src/main/java/com/mung/common/controller/ExceptionController.java
@@ -1,65 +1,0 @@
-package com.mung.common.controller;
-
-import com.mung.common.exception.CommonException;
-import com.mung.common.response.ErrorResponse;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.FieldError;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.*;
-
-@Slf4j
-@RestControllerAdvice
-public class ExceptionController {
-
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    @ResponseBody
-    public ErrorResponse inValidExceptionHandler(MethodArgumentNotValidException e) {
-
-        ErrorResponse errorResponse = ErrorResponse.builder()
-                .code("400")
-                .message("잘못된 요청입니다.")
-                .build();
-
-        for (FieldError fieldError : e.getFieldErrors()) {
-            errorResponse.addValidation(fieldError.getField(), fieldError.getDefaultMessage());
-        }
-
-        return errorResponse;
-    }
-
-    @ExceptionHandler(CommonException.class)
-    @ResponseBody
-    public ResponseEntity<ErrorResponse> commonException(CommonException e) {
-
-        int statusCode = e.getStatusCode();
-
-        ErrorResponse body = ErrorResponse.builder()
-                .code(String.valueOf(statusCode))
-                .message(e.getMessage())
-                .validation(e.getValidation())
-                .build();
-
-        ResponseEntity<ErrorResponse> response = ResponseEntity.status(statusCode)
-                .body(body);
-
-        return response;
-    }
-
-//    @ExceptionHandler(Exception.class)
-//    @ResponseBody
-//    public ResponseEntity<ErrorResponse> handleException(Exception e) {
-//
-//        log.error(e.getMessage());
-//
-//        ErrorResponse body = ErrorResponse.builder()
-//                .code(String.valueOf(500))
-//                .message("INTERNAL_SERVER_ERROR")
-//                .build();
-//
-//        return ResponseEntity.status(500).body(body);
-//    }
-
-}
