--- common/src/main/java/com/mung/common/domain/Address.java
@@ -1,19 +1,0 @@
-package com.mung.common.domain;
-
-import jakarta.persistence.Embeddable;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Embeddable
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class Address {
-
-    private String zipcode;
-    private String city;
-    private String street;
-
-}
