--- common/src/main/java/com/mung/common/domain/Role.java
@@ -1,18 +1,0 @@
-package com.mung.common.domain;
-
-import lombok.Getter;
-
-@Getter
-public enum Role {
-
-    USER("ROLE_USER"),
-    COMP("ROLE_COMP"),
-    ADMIN("ROLE_ADMIN");
-
-    private String full;
-
-    private Role(String full) {
-        this.full = full;
-    }
-
-}
