--- common/src/main/java/com/mung/common/exception/Unauthorized.java
@@ -1,20 +1,0 @@
-package com.mung.common.exception;
-
-public class Unauthorized extends CommonException {
-
-    private static final String MESSAGE = "인증이 필요 합니다.";
-
-    public Unauthorized() {
-        super(MESSAGE);
-    }
-
-    public Unauthorized(Throwable cause) {
-        super(MESSAGE, cause);
-    }
-
-    @Override
-    public int getStatusCode() {
-        return 401;
-    }
-
-}
