--- member/src/main/java/com/mung/member/MemberApplication.java
@@ -1,27 +1,0 @@
-package com.mung.member;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.context.annotation.Bean;
-import org.springframework.data.domain.AuditorAware;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-import java.util.Optional;
-import java.util.UUID;
-
-@SpringBootApplication
-@EnableCaching
-@EnableJpaAuditing
-public class MemberApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(MemberApplication.class, args);
-    }
-
-    @Bean
-    public AuditorAware<String> auditorProvider() {
-        return () -> Optional.of(UUID.randomUUID().toString()); // 수정자
-    }
-
-}
