--- member/src/main/java/com/mung/member/config/CommonBeanConfig.java
@@ -1,12 +1,0 @@
-package com.mung.member.config;
-
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@EnableAutoConfiguration
-@ComponentScan("com.mung.common")
-public class CommonBeanConfig {
-
-}
