+++ member/src/main/java/com/mung/member/domain/Address.java
@@ -1,0 +1,19 @@
+package com.mung.member.domain;
+
+import jakarta.persistence.Embeddable;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Embeddable
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class Address {
+
+    private String zipcode;
+    private String city;
+    private String street;
+
+}
