+++ member/src/main/java/com/mung/member/domain/Role.java
@@ -1,0 +1,18 @@
+package com.mung.member.domain;
+
+import lombok.Getter;
+
+@Getter
+public enum Role {
+
+    USER("ROLE_USER"),
+    COMP("ROLE_COMP"),
+    ADMIN("ROLE_ADMIN");
+
+    private String full;
+
+    private Role(String full) {
+        this.full = full;
+    }
+
+}
