+++ member/src/main/resources/application.yml
@@ -1,25 +1,5 @@
-spring:
-  datasource:
-    driver-class-name: com.mysql.cj.jdbc.Driver
-    url: jdbc:mysql://localhost:3306/member?serverTimezone=Asia/Seoul&characterEncoding=UTF-8
-    username: root
-    password: root
+#spring:
+#  redis:
+#    host: localhost
+#    port: 6379
 
-  jpa:
-    hibernate:
-      ddl-auto: update
-    properties:
-      hibernate:
-        format_sql: true
-    database-platform: org.hibernate.dialect.MySQLDialect
-
-  redis:
-    host: localhost
-    port: 6379
-
-logging.level:
-  org.hibernate.SQL: debug
-  # org.hibernate.type: trace
-
-server:
-  port: 8081
