+++ stock/build.gradle
@@ -1,3 +1,8 @@
 dependencies {
     implementation project(':common')
+
+}
+
+bootJar {
+    enabled = false
 }
