--- stock/src/main/java/com/mung/stock/StockApplication.java
@@ -1,13 +1,0 @@
-package com.mung.stock;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class StockApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(StockApplication.class, args);
-    }
-
-}
