--- stock/src/main/resources/application.yml
@@ -1,21 +1,0 @@
-spring:
-  datasource:
-    driver-class-name: com.mysql.cj.jdbc.Driver
-    url: jdbc:mysql://localhost:3306/stock?serverTimezone=Asia/Seoul&characterEncoding=UTF-8
-    username: root
-    password: root
-
-  jpa:
-    hibernate:
-      ddl-auto: update
-    properties:
-      hibernate:
-        format_sql: true
-    database-platform: org.hibernate.dialect.MySQLDialect
-
-logging.level:
-  org.hibernate.SQL: debug
-  # org.hibernate.type: trace
-
-server:
-  port: 8084
